deleted old Scheduler.js, refactored NEW_Scheduler.js to Scheduler.js
